+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Allow creating class names with no package (remove auto ns prefixing)
 * Add support for custom index objects via the `LocalIndex` protocol
 * Generate default constructor for :abstract classes too if one not provided
+* Support passing parameter and return types separately when defining methods
 
 ## 0.5.4 (2022-04-14)